# comsapi/app/src/docs/v1.api-spec.yaml
...
openapi: 3.0.2
info:
  version: 1.0.0
  title: Common Object Management Service (COMS)
  description: A microservice for managing access control to S3 Objects
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: NR Common Service Showcase
    email: NR.CommonServiceShowcase@gov.bc.ca
servers:
  - url: /api/v1
    description: This Server
security:
  - BasicAuth: []
  - BearerAuth: []
    OpenID: []
...
paths:
  /bucket:
    put:
      summary: Creates a bucket
      description: >-
        Creates a bucket record. Bucket should exist in S3. If the set of
        `bucket`, `endpoint` and `key` match an existing record, the user will
        be added to that existing bucket with full permissions instead of
        generating a new bucket record.
      operationId: createBucket
      tags:
        - Bucket
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request-CreateBucket'
      responses:
        '201':
          description: Returns inserted DB bucket record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DB-Bucket'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
    get:
      summary: Search for buckets
      description: >-
        Returns a list of buckets matching all search criteria across all bucket
        records
      operationId: searchBuckets
      tags:
        - Bucket
      parameters:
        - $ref: '#/components/parameters/Query-BucketId'
        - $ref: '#/components/parameters/Query-Active'
        - $ref: '#/components/parameters/Query-Key'
        - $ref: '#/components/parameters/Query-BucketName'
      responses:
        '200':
          description: Returns an array of buckets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DB-Bucket'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
  /bucket/{bucketId}:
    head:
      summary: Checks if a bucket exists in S3
      description: Returns either a 204 or 403 if user lacks permissions
      operationId: headBucket
      tags:
        - Bucket
      parameters:
        - $ref: '#/components/parameters/Path-BucketId'
      responses:
        '204':
          description: Returns object headers
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
    get:
      summary: Returns a bucket
      description: Returns a bucket record based on bucketId
      operationId: readBucket
      tags:
        - Bucket
      parameters:
        - $ref: '#/components/parameters/Path-BucketId'
      responses:
        '200':
          description: Returns bucket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DB-Bucket'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
    patch:
      summary: Updates a bucket
      description: Updates the bucket record
      operationId: updateBucket
      tags:
        - Bucket
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Request-UpdateBucket'
      parameters:
        - $ref: '#/components/parameters/Path-BucketId'
      responses:
        '200':
          description: Returns updated bucket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DB-Bucket'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
    delete:
      summary: Deletes a bucket
      description: >-
        Deletes the bucket record based on bucketId. This request does not
        dispatch an S3 operation to request the deletion of the associated
        bucket.
      operationId: deleteBucket
      tags:
        - Bucket
      parameters:
        - $ref: '#/components/parameters/Path-BucketId'
      responses:
        '204':
          description: Returns no content success
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
...
  /permission/bucket:
    get:
      summary: Searches for bucket permissions
      description: >-
        Returns an array of buckets and their permissions meeting the filtering
        parameters provided. Use `objectPerms=true` in conjunction with a userId
        to get an extended bucket list factoring in object inheritance.
      operationId: bucketSearchPermissions
      tags:
        - Permission
      parameters:
        - $ref: '#/components/parameters/Query-BucketId'
        - $ref: '#/components/parameters/Query-UserId'
        - $ref: '#/components/parameters/Query-PermCode'
        - $ref: '#/components/parameters/Query-ObjectPerms'
      responses:
        '200':
          description: >-
            Returns an array of bucketIds and permissions that match the
            provided parameters
          content:
            application/json:
              schema:
                type: array
                items:
                  anyOf:
                    - $ref: '#/components/schemas/Response-BucketPerms'
                    - $ref: '#/components/schemas/Response-BucketPermsWithObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
  /permission/bucket/{bucketId}:
    get:
      summary: Returns the bucket permissions
      description: >-
        Returns an array of permissions for a specific bucket meeting the
        filtering parameters provided
      operationId: bucketListPermissions
      tags:
        - Permission
      parameters:
        - $ref: '#/components/parameters/Query-UserId'
        - $ref: '#/components/parameters/Path-BucketId'
        - $ref: '#/components/parameters/Query-PermCode'
      responses:
        '200':
          description: >-
            Returns an array of bucketId/userId/permCode triplets that match the
            provided parameters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DB-BucketPermission'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
    put:
      summary: Grants bucket permissions to users
      description: >-
        Adds permissions for a specific bucket to users by specifying an
        arbitrary array of permCode and user tuples. This is an idempotent
        operation, so users that already have a requested permission will remain
        unaffected. Only permissions successfully added to COMS will appear in
        the response.
      operationId: bucketAddPermissions
      tags:
        - Permission
      parameters:
        - $ref: '#/components/parameters/Path-BucketId'
      requestBody:
        description: >-
          An array of bucket permissions, each containing a `userId` and
          `permCode`
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Request-PermissionTuple'
      responses:
        '201':
          description: Returns an array of added permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DB-BucketPermission'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
    delete:
      summary: Deletes permissions for a bucket
      description: >-
        Removes permissions for a specific bucket by optionally specifying a set
        of users and subset of permissions to revoke. This is an idempotent
        operation, so users that already lack the specified permission(s) will
        remain unaffected. Only permissions successfully removed from COMS will
        appear in the response. WARNING: Specifying no parameters will delete
        all permissions associated with a bucket; it is possible to lock
        yourself out of your own bucket!
      operationId: bucketRemovePermissions
      tags:
        - Permission
      parameters:
        - $ref: '#/components/parameters/Path-BucketId'
        - $ref: '#/components/parameters/Query-UserId'
        - $ref: '#/components/parameters/Query-PermCode'
      responses:
        '200':
          description: Returns an array of deleted permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DB-BucketPermission'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
...
components:
...
  parameters:
...
    Path-BucketId:
      in: path
      name: bucketId
      description: Uuid of a bucket
      required: true
      schema:
        type: string
        format: uuid
        example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
...
    Query-Active:
      in: query
      name: active
      description: Boolean on active status
      schema:
        type: boolean
        example: true
...
    Query-BucketId:
      in: query
      name: bucketId
      description: Uuid or array of uuids representing the bucket
      schema:
        oneOf:
          - type: string
            format: uuid
            example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
          - type: array
            items:
              type: string
              format: uuid
              example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
    Query-BucketName:
      in: query
      name: name
      description: A display name given to the bucket on creation
      schema:
        type: string
        example: Environmental Documents
...
    Query-Key:
      in: query
      name: key
      description: The prefix given to objects in the bucket
      schema:
        type: string
        example: coms/env
...
    Query-ObjectPerms:
      in: query
      name: objectPerms
      description: >-
        Boolean representing whether or not to include buckets containing
        objects with matching permissions
      schema:
        type: boolean
        example: true
...
    Query-PermCode:
      in: query
      name: permCode
      description: The permission type
      schema:
        oneOf:
          - $ref: '#/components/schemas/PermCode'
          - type: array
            items:
              $ref: '#/components/schemas/PermCode'
...
    Query-UserId:
      in: query
      name: userId
      description: Uuid or array of uuids representing the user
      schema:
        oneOf:
          - type: string
            format: uuid
            example: 00000000-0000-0000-0000-000000000000
          - type: array
            items:
              type: string
              format: uuid
              example: 00000000-0000-0000-0000-000000000000
...
  schemas:
    DB-Bucket:
      title: DB Bucket
      type: object
      allOf:
        - type: object
          required:
            - accessKeyId
            - active
            - bucket
            - bucketId
            - bucketName
            - endpoint
            - key
            - region
            - secretAccessKey
          properties:
            accessKeyId:
              type: string
              description: >-
                S3 account Access Key Id, similar to a username. This value will
                be redacted from the API response.
              example: REDACTED
            active:
              type: boolean
              description: Determines whether this bucket is considered active in COMS
              default: true
              example: true
            bucket:
              type: string
              description: The ID of a bucket
              example: ezakyp
            bucketName:
              type: string
              description: A human-readable label or title
              example: Geospatial Maps
            bucketId:
              type: string
              description: The primary identifier for this bucket
              format: uuid
              example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
            endpoint:
              type: string
              description: A valid RFC 3986 formatted S3 Service URI endpoint
              maxLength: 255
              example: https://my-objectstore.ca
            key:
              description: The prefix given to objects in the bucket
              type: string
              example: coms/env
              maxLength: 255
            region:
              type: string
              description: >-
                The geographical zone defined by the S3 service where your
                bucket is located
              example: ca-central-1
            secretAccessKey:
              type: string
              description: >-
                S3 account Secret Access Key, similar to a password. This value
                will be redacted from the API response.
              example: REDACTED
        - $ref: '#/components/schemas/DB-TimestampUserData'
    DB-BucketPermission:
      title: DB Bucket Permission
      type: object
      allOf:
        - type: object
          required:
            - id
            - bucketId
            - userId
            - permCode
          properties:
            id:
              type: string
              format: uuid
              description: The unique identifier for this permission tuple
              example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
            bucketId:
              type: string
              format: uuid
              description: The unique identifier for the bucket
              example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
            userId:
              type: string
              format: uuid
              description: The unique identifier of the user
              example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
            permCode:
              $ref: '#/components/schemas/PermCode'
        - $ref: '#/components/schemas/DB-TimestampUserData'
...
    DB-TimestampUserData:
      title: DB Timestamp User Data
      type: object
      required:
        - createdBy
        - createdAt
      properties:
        createdBy:
          type: string
          format: uuid
          description: >-
            The subject id of the current user if request was authenticated with
            a Bearer token (ex. JWT), or a 'nil' uuid if request was
            authenticated via Basic auth
          example: 72cf13d7-aef7-4ad5-8ddd-0ff23eefb442
        createdAt:
          type: string
          format: date-time
          description: Time when this record was created
          example: '2022-03-11T23:19:16.343Z'
        updatedBy:
          type: string
          format: uuid
          description: >-
            The subject id of the current user if request was authenticated with
            a Bearer token (ex. JWT), or a 'nil' uuid if request was
            authenticated via Basic auth
          default: null
          example: 72cf13d7-aef7-4ad5-8ddd-0ff23eefb442
        updatedAt:
          type: string
          format: date-time
          description: Time when this record was last updated
          example: '2022-03-11T23:19:16.343Z'
...
    PermCode:
      title: Permission Code
      type: string
      description: Permission code/type
      enum:
        - CREATE
        - READ
        - UPDATE
        - DELETE
        - MANAGE
      example: UPDATE
    Request-CreateBucket:
      title: Request Create Bucket
      type: object
      required:
        - accessKeyId
        - bucket
        - bucketName
        - endpoint
        - secretAccessKey
      allOf:
        - $ref: '#/components/schemas/Request-UpdateBucket'
        - type: object
          properties:
            key:
              description: >-
                The prefix given to objects in the bucket. Defaults to "/" if
                undefined.
              type: string
              default: /
              example: coms/env
    Request-UpdateBucket:
      title: Request Update Bucket
      type: object
      properties:
        accessKeyId:
          type: string
          description: >-
            S3 account Access Key Id, similar to a username. This value will be
            redacted from the API response.
          example: REDACTED
        active:
          type: boolean
          description: Whether this bucket is considered active in COMS
          default: true
          example: true
        bucket:
          type: string
          description: The ID of a bucket
          example: ezakyp
        bucketName:
          type: string
          description: A human-readable label or title
          example: Geospatial Maps
        endpoint:
          type: string
          description: A valid RFC 3986 formatted S3 Service URI endpoint
          maxLength: 255
          example: https://my-objectstore.ca
        region:
          type: string
          description: >-
            The geographical zone defined by the S3 service where your bucket is
            located
          example: ca-central-1
        secretAccessKey:
          type: string
          description: >-
            S3 account Secret Access Key, similar to a password. This value will
            be redacted from the API response.
          example: REDACTED
    Request-PermissionTuple:
      title: Response Permission Tuple
      type: object
      required:
        - userId
        - permCode
      properties:
        permCode:
          $ref: '#/components/schemas/PermCode'
        userId:
          type: string
          description: The primary identifier for a user
          format: uuid
          example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
...
    Response-BucketPerms:
      title: Response Buckets with bucket-level permissions
      type: object
      properties:
        bucketId:
          type: string
          format: uuid
          description: >-
            The unique identifier for the bucket. When `objectPerms=false` only
            buckets with bucket-level permissions are returned.
          example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/DB-BucketPermission'
    Response-BucketPermsWithObject:
      title: Response Buckets containing objects with object-level permissions
      description: >-
        When `objectPerms=true` response will include buckets containing objects
        with matching permissions.
      type: object
      properties:
        bucketId:
          type: string
          format: uuid
          description: >-
            The unique identifier for the bucket. When `objectPerms=true`
            response will include buckets containing objects with matching
            permissions.
          example: bf246e31-c807-496c-bc93-cd8bc2f1897f
        permissions:
          type: array
          items:
            type: object
          description: Array will be empty.
          example: []
    Response-Conflict:
      title: Response Conflict
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            status:
              example: 409
            title:
              example: Conflict
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/409
...
    Response-Error:
      title: Response Error
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            status:
              example: 500
            title:
              example: Internal Server Error
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/500
    Response-Forbidden:
      title: Response Forbidden
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/403
            title:
              example: Forbidden
            status:
              example: 403
            detail:
              example: User lacks permission to complete this action
...
    Response-Problem:
      title: Response Problem
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          description: What type of problem, link to explanation of problem
        title:
          type: string
          description: Title of problem, generally the HTTP Status Code description
        status:
          type: string
          description: The HTTP Status code
        detail:
          type: string
          description: >-
            A short, human-readable explanation specific to this occurrence of
            the problem
        instance:
          type: string
          description: >-
            A URI reference that identifies the specific occurrence of the
            problem
    Response-Unauthorized:
      title: Response Unauthorized
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/401
            title:
              example: Unauthorized
            status:
              example: 401
            detail:
              example: Invalid authorization credentials
...
  responses:
...
    Conflict:
      description: Conflict (Request conflicts with server state)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Conflict'
...
    Error:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Error'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Forbidden'
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Unauthorized'
...
  securitySchemes:
    BasicAuth:
      type: http
      description: Basic auth via environment defined username and password
      scheme: basic
    BearerAuth:
      type: http
      description: Bearer token auth using an OIDC issued JWT token
      scheme: bearer
      bearerFormat: JWT
    OpenID:
      type: openIdConnect
      description: OpenID Connect endpoint for acquiring JWT tokens
      openIdConnectUrl: >-
        https://logonproxy.gov.bc.ca/auth/realms/your-realm-name/.well-known/openid-configuration